// src/app/modules/cpassapi/model/ricercaEvasioni.ts
/*
* SPDX-FileCopyrightText: Copyright 2019 - 2020 | CSI Piemonte
* SPDX-License-Identifier: EUPL-1.2
*/
/**
 * Contabilità passiva
 * API per il backend della suite di contabilità passiva.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { DestinatarioEvasione } from './destinatarioEvasione';
import { Impegno } from './impegno';
import { OggettiSpesa } from './oggettiSpesa';
import { Subimpegno } from './subimpegno';
import { TestataEvasione } from './testataEvasione';


export interface RicercaEvasioni { 
    annoEvasioneDa?: number;
    numeroEvasioneDa?: number;
    annoEvasioneA?: number;
    numeroEvasioneA?: number;
    dataInserimentoDa?: Date;
    dataInserimentoA?: Date;
    annoOrdineDa?: number;
    numeroOrdineDa?: number;
    annoOrdineA?: number;
    numeroOrdineA?: number;
    dataEmissioneDa?: Date;
    dataEmissioneA?: Date;
    annoProvvedimento?: number;
    numeroProvvedimento?: number;
    testataEvasione?: TestataEvasione;
    destinatarioEvasione?: DestinatarioEvasione;
    impegno?: Impegno;
    subimpegno?: Subimpegno;
    oggettiSpesa?: OggettiSpesa;
}